backend/pyproject.toml
[project]
name = "prontia"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "kota_na_", email = "ad.maiora.1@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-cosmos>=4.9.0",
    "fastapi[standard]>=0.117.1",
    "openai>=1.108.2",
    "uuid6>=2025.0.1",
]

[project.scripts]
prontia = "prontia.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.1",
]

[[tool.mypy.overrides]]
module = [
    'prontia.*',
]
ignore_missing_imports = true
